refactor: ResultProcessor 의존성 구현 패키지로 변경
